automation/requirements.txt
azureml-sdk==1.52.0
mlflow==2.5.0
nltk==3.7
numpy==1.24.3
pandas==1.2.5
python-dotenv==1.0.0
scikit-learn==1.1.3
